Clear env. variable ERL_FLAGS before obtaining the erlang version
In cases where the name of a running node is set through
the environment variable ERL_FLAGS (e.q. ERL_FLAGS="-sname test"),
the command to fetch the erlang version will try to start a node
with the same configs and fail. A local clear of ERL_FLAGS for that
command will prevent this behavior.