Tweak Erlang/OTP version determination for releases >= 17
As of OTP release 17, the OTP release number corresponds to the major
part of the OTP version and the actual current OTP version can be read
from the text file
<OTP installation root>/releases/<OTP release number>/OTP_VERSION or
<OTP source root>/OTP_VERSION. As a fallback for prior releases or in
case of an error, the former behaviour will be used. It also affects
the creation of PLTs insofar that for each specific OTP release its
own PLT including the full version will be created. In addition
a 'otp_version' convenient target was added.